Support multi-codepoint emojis
Some emojis are composed of several codepoints, for example the flag
emojis, or the family etc emojis using the Zero-width joiner character.

We need to change the signature of the cdawg_match function to support
them. Instead of returning a pair of a codepoint and a match length, we
return a null-terminated list of codepoints along with the match length.

Assumes the codepoints defined in mapping.json are dash separated like
in this example:

    ":man-woman-boy-boy:": "1F468-200D-1F469-200D-1F466-200D-1F466"